Added s3 download and random image functions
List functions are now private because they shouldn't need to be used
publicly, and making them private simplifies the interface. Started
fleshing out s3 functions for downloading and randomizing images. For
the future, see if it's possible to use io streams to avoid having to
save/open/save/upload the file in the process of manipulating it.